+++ tests/testthat/test-type.R
@@ -175,3 +175,13 @@ test_that("xportr_type: error when metadata is not set", {
     regexp = "Metadata must be set with `metadata` or `xportr_metadata\\(\\)`"
   )
 })
+
+test_that("xportr_type: Gets warning when metadata has multiple rows with same variable", {
+  # This test uses the (2) functions below to reduce code duplication
+  # All `expect_*` are being called inside the functions
+  #
+  # Checks that message appears when xportr.domain_name is invalid
+  multiple_vars_in_spec_helper(xportr_type)
+  # Checks that message doesn't appear when xportr.domain_name is valid
+  multiple_vars_in_spec_helper2(xportr_type)
+})