+++ NEWS.md
@@ -10,6 +10,7 @@
 * `xpt_validate` updated to accept iso8601 date formats. (#76)
 * Added function `xportr_metadata()` to explicitly set metadata at the start of a pipeline (#44)
 * Metadata order columns are now coerced to numeric by default in `xportr_order()` to prevent character sorting (#149)
+* Message is shown on `xportr_*` functions when the metadata being used has multiple variables with the same name in the same domain (#128)
 
 ## Documentation